TST: Merge 2d and 3d tests into a single file
